Move grid to a subfolder & rename grid tests
