+++ lib/screens/home/home.dart
@@ -4,7 +4,7 @@ import 'package:flutter_redux/flutter_redux.dart';
 
 import 'package:learning_compass_exp/store/app_state.dart';
 import 'package:learning_compass_exp/screens/home/widgets/flower/flower_menu.dart';
-import 'package:learning_compass_exp/screens/home/widgets/grid.dart';
+import 'package:learning_compass_exp/screens/home/widgets/grid/grid.dart';
 
 // Use this file as a starting-point to build the contents of the page.
 // Fill it with Widgets from the /home/widgets directory
