+++ test/widget_tests/grid_item_widget_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:learning_compass_exp/screens/home/widgets/grid_item.dart';
+import 'package:learning_compass_exp/screens/home/widgets/grid/grid_item.dart';
 import 'package:learning_compass_exp/data/models/petal.dart';
 
 import 'package:learning_compass_exp/data/constants/PETAL_CONSTANTS.dart';
