+++ test/widget_tests/grid_widget_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:learning_compass_exp/screens/home/widgets/grid.dart';
-import 'package:learning_compass_exp/screens/home/widgets/grid_item.dart';
+import 'package:learning_compass_exp/screens/home/widgets/grid/grid.dart';
+import 'package:learning_compass_exp/screens/home/widgets/grid/grid_item.dart';
 
 import 'package:learning_compass_exp/data/models/petal.dart';
